Remove code to install CRD
