+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "kube-example-cnat"
+name = "cnat"
 version = "0.1.0"
 description = "Cloud Native at. Runs `command` at `schedule` by creating a Pod."
 keywords = ["kubernetes", "k8s", "kube", "example"]
@@ -9,13 +9,9 @@ readme = "README.md"
 authors = ["kazk <kazk.dev@gmail.com>"]
 license = "Apache-2.0"
 edition = "2018"
-
-[[bin]]
-name = "cnat"
-path = "src/main.rs"
+default-run = "cnat"
 
 [dependencies]
-argh = "0.1.4"
 chrono = { version = "0.4", features = ["serde"] }
 futures = "0.3.5"
 kube = { version = "0.50.1", features = ["derive"] }
