+++ README.md
@@ -21,8 +21,6 @@ Example Kubernetes controller [programming-kubernetes/cnat] in Rust using [kube]
 - Changed to use `apiextensions.k8s.io/v1` and added the required `schema`
   - The schema is generated by `kube` ([clux/kube-rs#348])
   - `schedule` is validated as `date-time`
-- Added option to output CRD (`cnat -g/--gen-crd`)
-- Added option to automatically install the current CRD if not already installed (`cnat -a/--auto-install`)
 
 <details>
 <summary>Generated CRD</summary>
@@ -92,7 +90,8 @@ spec:
 
 ```bash
 # Create CRD
-cargo run -- --gen-crd | kubectl apply -f -
+kubectl apply -f k8s/ats.yml
+kubectl wait --for=condition=NamesAccepted crd/ats.example.kazk.dev
 
 # Start controller
 cargo run
@@ -108,7 +107,8 @@ kubectl apply -f k8s/example.yml
 kubectl apply -f k8s/rbac.yml
 
 # Create CRD
-docker run --rm ghcr.io/kazk/cnat:latest --gen-crd | kubectl apply -f -
+kubectl apply -f k8s/ats.yml
+kubectl wait --for=condition=NamesAccepted crd/ats.example.kazk.dev
 
 # Deploy controller
 kubectl apply -f k8s/deployment.yml
@@ -131,7 +131,7 @@ Building on Arch Linux:
 - `client-go`: 31MB (23MB with `-ldflags "-s -w"`)
 - `kubebuilder`: TBD requires unknown version of [KubeBuilder]
 - `operator-sdk`: TBD requires unknown version of [Operator SDK]
-- `kube`: 11MB (8.2MB with `RUSTFLAGS='-C link-arg=-s'`)
+- `kube`: 8.3MB (5.8MB with `RUSTFLAGS='-C link-arg=-s'`)
 
 ### Lines of Code
 
@@ -140,7 +140,7 @@ Building on Arch Linux:
 | client-go     |    29    |   2161  |   1212  |       658 |     291 |
 | kubebuilder   |    16    |   1013  |    529  |       357 |     127 |
 | operator-sdk  |    13    |    714  |    488  |       134 |      92 |
-| kube          |     4    |    329  |    281  |        13 |      35 |
+| kube          |     6    |    274  |    230  |        15 |      29 |
 
 
 - `client-go`: `tokei -e cnat-client-go/vendor -e cnat-client-go/hack -t Go cnat-client-go`
