+++ k8s/ats.yml
@@ -1,0 +1,56 @@
+---
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  name: ats.example.kazk.dev
+spec:
+  group: example.kazk.dev
+  names:
+    kind: At
+    plural: ats
+    shortNames: []
+    singular: at
+  scope: Namespaced
+  versions:
+    - additionalPrinterColumns: []
+      name: v1alpha1
+      schema:
+        openAPIV3Schema:
+          description: "Auto-generated derived type for AtSpec via `CustomResource`"
+          properties:
+            spec:
+              description: Spec for custom resource At.
+              properties:
+                command:
+                  items:
+                    type: string
+                  type: array
+                schedule:
+                  format: date-time
+                  type: string
+              required:
+                - command
+                - schedule
+              type: object
+            status:
+              description: Status for custom resource At.
+              nullable: true
+              properties:
+                phase:
+                  description: Describes the status of the scheduled command.
+                  enum:
+                    - Running
+                    - Done
+                  type: string
+              required:
+                - phase
+              type: object
+          required:
+            - spec
+          title: At
+          type: object
+      served: true
+      storage: true
+      subresources:
+        status: {}
+
