+++ k8s/rbac.yml
@@ -14,11 +14,6 @@ apiVersion: rbac.authorization.k8s.io/v1
 metadata:
   name: control-ats
 rules:
-  # Allow checking if CRD is installed
-  - apiGroups: ["apiextensions.k8s.io"]
-    resources: ["customresourcedefinitions"]
-    verbs:
-      - get
   - apiGroups: ["example.kazk.dev"]
     resources: ["ats", "ats/status"]
     verbs:
