+++ src/bin/crd.rs
@@ -1,0 +1,7 @@
+// Output CRD
+// cargo run --bin crd
+use cnat::At;
+
+fn main() {
+    println!("{}", serde_yaml::to_string(&At::crd()).unwrap())
+}
