+++ src/error.rs
@@ -1,7 +1,6 @@
 use chrono::ParseError;
 use kube::Error as KubeError;
 use serde_json::Error as JsonError;
-use serde_yaml::Error as YamlError;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
@@ -12,14 +11,8 @@ pub enum Error {
     ParseScheduleError(ParseError),
     #[error("JSON Error: {0}")]
     SerdeJsonError(JsonError),
-    #[error("YAML Error: {0}")]
-    SerdeYamlError(YamlError),
     #[error("Missing object key: {0}")]
     MissingObjectKey(&'static str),
-    #[error("Timed out: {0}")]
-    TimedOut(String),
-    #[error("Missing CRD")]
-    MissingCRD,
 }
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
@@ -39,11 +32,5 @@ impl From<ParseError> for Error {
 impl From<JsonError> for Error {
     fn from(error: JsonError) -> Self {
         Error::SerdeJsonError(error)
-    }
-}
-
-impl From<YamlError> for Error {
-    fn from(error: YamlError) -> Self {
-        Error::SerdeYamlError(error)
     }
 }
