+++ src/lib.rs
@@ -1,0 +1,7 @@
+mod controller;
+mod error;
+mod resource;
+
+pub use controller::run;
+pub use error::{Error, Result};
+pub use resource::At;
